+++ JaProjektFiltr/Filter/AssemblyFilter.cs
@@ -14,9 +14,9 @@ namespace JaProjektFiltr.Filter
         public AssemblyFilter(int bytesPerPixel,int startIndex, int endIndex): base( bytesPerPixel, startIndex, endIndex) { }
 
         [DllImport("C:\\Users\\Marcin\\Desktop\\Gradientowy-filtr-kierunkowy-wschodni\\x64\\Debug\\AsmProjekt.dll", EntryPoint = "AsmProc")]
-        private static extern void AsmProc(float[] pixels, float[] array255, int startIndex, int endIndex);
+        private static extern void AsmProc(byte[] pixels, float[] array255, int startIndex, int endIndex);
 
-        public override void ExecuteResult(float[] allPixels)
+        public override void ExecuteResult(byte[] allPixels)
         {
 
             AsmProc(allPixels, array255, _startIndex, _endIndex);
